+++ hello-world-scaffolded/.vscode/Invoke.ps1
@@ -1,0 +1,3 @@
+
+start powershell {./.vscode./CompileAllComponents.ps1}
+